feat(couchdb): support for linked documents in couchdb
This commit adds a method to the database connector which allows the retrieval of linked documents
from couchdb as described in http://docs.couchdb.org/en/2.1.0/ddocs/views/joins.html.

Read the documentation for the annotations LinkedDocument and LinkedDocuments as well as the
test cases in LinkedDocumentLoadingTest for more information about how it works.

Another source might be https://github.com/sw360/sw360portal/wiki/External-documents-with-CouchDB